third-task/pyproject.toml
[tool.poetry]
name = "svd"
version = "0.1.0"
description = ""
authors = ["raf-nr <nurmuhametovrafik04@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
